+++ apps/nextjs/paths.txt
@@ -1,12 +1,30 @@
-src/components/pages/inbox/Component.tsx
-src/components/pages/inbox/components/Activity/Activity.tsx
-src/components/pages/inbox/components/Activity/index.ts
-src/components/pages/inbox/components/Archive/Archive.tsx
-src/components/pages/inbox/components/Archive/index.ts
-src/components/pages/inbox/components/Header/Header.tsx
-src/components/pages/inbox/components/Header/index.ts
-src/components/pages/inbox/components/Header/Tabs.tsx
-src/components/pages/inbox/Container.tsx
-src/components/pages/inbox/Inbox.stories.tsx
-src/components/pages/inbox/index.ts
-src/components/pages/inbox/providers/Provider.tsx
+src/components/pages/my_tasks/Component.tsx
+src/components/pages/my_tasks/components/Board/Board.tsx
+src/components/pages/my_tasks/components/Board/index.ts
+src/components/pages/my_tasks/components/Board/SkeletonBoard/index.ts
+src/components/pages/my_tasks/components/Board/SkeletonBoard/SkeletonBoard.tsx
+src/components/pages/my_tasks/components/Board/SkeletonBoard/SkeletonBoardHeader.tsx
+src/components/pages/my_tasks/components/Calendar/Calendar.tsx
+src/components/pages/my_tasks/components/Calendar/index.ts
+src/components/pages/my_tasks/components/Calendar/SkeletonCalendar.tsx
+src/components/pages/my_tasks/components/Files/Files.tsx
+src/components/pages/my_tasks/components/Files/index.ts
+src/components/pages/my_tasks/components/Files/SkeletonFiles.tsx
+src/components/pages/my_tasks/components/Header/Avatar.tsx
+src/components/pages/my_tasks/components/Header/Header.tsx
+src/components/pages/my_tasks/components/Header/index.ts
+src/components/pages/my_tasks/components/Header/ShareButton.tsx
+src/components/pages/my_tasks/components/Header/SkeletonHeader.tsx
+src/components/pages/my_tasks/components/Header/Tabs.tsx
+src/components/pages/my_tasks/components/List/index.ts
+src/components/pages/my_tasks/components/List/List.tsx
+src/components/pages/my_tasks/components/List/SkeletonList/index.ts
+src/components/pages/my_tasks/components/List/SkeletonList/SkeletonListContent.tsx
+src/components/pages/my_tasks/components/List/SkeletonList/SkeletonListHeader.tsx
+src/components/pages/my_tasks/components/TasksHeader/index.ts
+src/components/pages/my_tasks/components/TasksHeader/SortMenu/index.ts
+src/components/pages/my_tasks/components/TasksHeader/SortMenu/SortMenu.tsx
+src/components/pages/my_tasks/Container.tsx
+src/components/pages/my_tasks/index.ts
+src/components/pages/my_tasks/MyTasks.stories.tsx
+src/components/pages/my_tasks/providers/Provider.tsx
