+++ apps/nextjs/src/app/my_tasks/layout.tsx
@@ -1,4 +1,4 @@
-import { Container } from '@/components/pages/my_tasks';
+import { Container } from '@/components/pages/my-tasks';
 
 export default async function MyTasksLayout() {
   return <Container />;
