+++ apps/nextjs/src/components/pages/my-tasks/components/board/index.ts
@@ -1,0 +1,1 @@
+export { Board } from './board';
