+++ apps/nextjs/src/components/pages/my-tasks/components/board/skeleton-board/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonBoardContent } from './skeleton-board';
+export { SkeletonBoardHeader } from './skeleton-board-header';
