+++ apps/nextjs/src/components/pages/my-tasks/components/calendar/index.ts
@@ -1,0 +1,1 @@
+export { Calendar } from './calendar';
