+++ apps/nextjs/src/components/pages/my-tasks/components/files/index.ts
@@ -1,0 +1,1 @@
+export { Files } from './files';
