+++ apps/nextjs/src/components/pages/my-tasks/components/list/index.ts
@@ -1,0 +1,1 @@
+export { List } from './list';
