+++ apps/nextjs/src/components/pages/my-tasks/components/list/skeleton-list/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonListContent } from './skeleton-list-content';
+export { SkeletonListHeader } from './skeleton-list-header';
