+++ apps/nextjs/src/components/pages/my-tasks/components/list/skeleton-list/skeleton-list-content.tsx
