+++ apps/nextjs/src/components/pages/my-tasks/components/tasks-header/index.ts
@@ -1,0 +1,1 @@
+export { SortMenu } from './sort-menu';
