+++ apps/nextjs/src/components/pages/my-tasks/components/tasks-header/sort-menu/index.ts
@@ -1,0 +1,1 @@
+export { SortMenu } from './sort-menu';
