--- apps/nextjs/src/components/pages/my_tasks/components/Board/SkeletonBoard/index.ts
@@ -1,2 +1,0 @@
-export { SkeletonBoardContent } from './SkeletonBoard';
-export { SkeletonBoardHeader } from './SkeletonBoardHeader';
