--- apps/nextjs/src/components/pages/my_tasks/components/Board/index.ts
@@ -1,1 +1,0 @@
-export { Board } from './Board';
