--- apps/nextjs/src/components/pages/my_tasks/components/Calendar/index.ts
@@ -1,1 +1,0 @@
-export { Calendar } from './Calendar';
