--- apps/nextjs/src/components/pages/my_tasks/components/Files/index.ts
@@ -1,1 +1,0 @@
-export { Files } from './Files';
