--- apps/nextjs/src/components/pages/my_tasks/components/Header/index.ts
@@ -1,1 +1,0 @@
-export { Header } from './Header';
