--- apps/nextjs/src/components/pages/my_tasks/components/List/SkeletonList/index.ts
@@ -1,2 +1,0 @@
-export { SkeletonListContent } from './SkeletonListContent';
-export { SkeletonListHeader } from './SkeletonListHeader';
