--- apps/nextjs/src/components/pages/my_tasks/components/List/index.ts
@@ -1,1 +1,0 @@
-export { List } from './List';
