--- apps/nextjs/src/components/pages/my_tasks/components/TasksHeader/SortMenu/index.ts
@@ -1,1 +1,0 @@
-export { SortMenu } from './SortMenu';
