--- apps/nextjs/src/components/pages/my_tasks/index.ts
@@ -1,1 +1,0 @@
-export { Container } from './Container';
